+++ eelbrain/test.py
@@ -4,7 +4,8 @@ __test__ = False
 from ._stats.glm import anova, ANOVA
 from ._stats.test import (
     Correlation,
-    TTest1Sample, TTestInd, TTestRel, pairwise, ttest, correlations,
+    TTest1Sample, TTestInd, TTestRel, MannWhitneyU, WilcoxonSignedRank,
+    pairwise, ttest, correlations,
     bootstrap_pairwise, lilliefors,
     pairwise_correlations,
 )